refactor: retarget storage container cache owner
